Introduce useOpenings hook
