+++ packages/ui/src/app/pages/WorkingGroups/WorkingGroupsOpenings.tsx
@@ -1,12 +1,15 @@
 import React from 'react'
 import { useHistory } from 'react-router-dom'
 
-import { ContentWithSidepanel } from '../../../common/components/page/PageContent'
+import { ContentWithSidepanel, MainPanel, SidePanel } from '../../../common/components/page/PageContent'
 import { Statistics } from '../../../common/components/statistics/Stats'
+import { OpeningsList } from '../../../working-groups/components/OpeningsList'
+import { useOpenings } from '../../../working-groups/hooks/useWorkingGroups'
 import { AppPage } from '../../components/AppPage'
 
 export const WorkingGroupsOpenings = () => {
   const history = useHistory()
+  const openings = useOpenings()
 
   const tabs = [
     { title: 'Openings', active: true, onClick: () => history.push('/working-groups') },
@@ -21,22 +24,26 @@ export const WorkingGroupsOpenings = () => {
   return (
     <AppPage pageTitle="Working Groups" crumbs={crumbs} tabs={tabs}>
       <ContentWithSidepanel>
-        <Statistics
-          stats={[
-            {
-              title: 'MyRoles',
-              value: 5,
-            },
-            {
-              title: 'Currently staking',
-              value: 5,
-            },
-            {
-              title: 'Earned in past',
-              value: 5,
-            },
-          ]}
-        />
+        <MainPanel>
+          <Statistics
+            stats={[
+              {
+                title: 'MyRoles',
+                value: 5,
+              },
+              {
+                title: 'Currently staking',
+                value: 5,
+              },
+              {
+                title: 'Earned in past',
+                value: 5,
+              },
+            ]}
+          />
+          <OpeningsList openings={openings} />
+        </MainPanel>
+        <SidePanel>Activities</SidePanel>
       </ContentWithSidepanel>
     </AppPage>
   )
