+++ packages/ui/src/working-groups/hooks/useWorkingGroups.ts
@@ -1,5 +1,8 @@
+import BN from 'bn.js'
+
 import { useGetWorkingGroupsQuery } from '../queries'
-import { WorkingGroup } from '../types'
+import { WorkingGroup, WorkingGroupOpening } from '../types'
+
 interface UseWorkingGroups {
   isLoading: boolean
   groups: WorkingGroup[]
@@ -12,4 +15,29 @@ export const useWorkingGroups = (): UseWorkingGroups => {
     isLoading: loading,
     groups,
   }
+}
+
+export const useOpenings = () => {
+  const openings: WorkingGroupOpening[] = [
+    {
+      id: '123',
+      title: 'Storage working group leader',
+      duration: [123, 'days'],
+      type: 'LEADER',
+      reward: { value: new BN(1000), interval: 3600 },
+      applicants: { current: 2, total: 10 },
+      hiring: { current: 0, total: 1 },
+    },
+    {
+      id: '221',
+      title: 'Storage working group worker',
+      duration: [12, 'days'],
+      type: 'LEADER',
+      reward: { value: new BN(1000), interval: 3600 },
+      applicants: { current: 2, total: 10 },
+      hiring: { current: 0, total: 1 },
+    },
+  ]
+
+  return openings
 }
